unify groundedness start
